requirements.txt
keras
Pillow
wandb
numpy